refactor(ui): flatten `reports/` dir
- same as previous commits
- notably `workflows-to-chart-data.ts` isn't a component, it's a util module
