+++ ui/src/app/reports/index.ts
@@ -1,4 +1,4 @@
-import {ReportsContainer} from './components/report-container';
+import {ReportsContainer} from './report-container';
 
 export default {
     component: ReportsContainer
