+++ ui/src/app/reports/reports-filters.scss
