+++ ui/src/app/reports/reports-filters.tsx
@@ -1,10 +1,10 @@
 import * as React from 'react';
 
-import {NODE_PHASE} from '../../../models';
-import {DataLoaderDropdown} from '../../shared/components/data-loader-dropdown';
-import {NamespaceFilter} from '../../shared/components/namespace-filter';
-import {TagsInput} from '../../shared/components/tags-input/tags-input';
-import {services} from '../../shared/services';
+import {NODE_PHASE} from '../../models';
+import {DataLoaderDropdown} from '../shared/components/data-loader-dropdown';
+import {NamespaceFilter} from '../shared/components/namespace-filter';
+import {TagsInput} from '../shared/components/tags-input/tags-input';
+import {services} from '../shared/services';
 
 import './reports-filters.scss';
